KVM: PPC: Associate IOMMU group with guest copy of TCE table
The existing in-kernel TCE table for emulated devices contains
guest physical addresses which are accesses by emulated devices.
Since we need to keep this information for VFIO devices too
in order to implement H_GET_TCE, we are reusing it.

This adds IOMMU group list to kvmppc_spapr_tce_table. Each group
will have or not have an iommu_table pointer:
1) IODA2 case - IOMMU tables will be most likely shared so only one
item in the list will have non-null @tbl;
2) IODA1/P5IOC2 - IOMMU tables won't be shared so every item in
the list will have non-null @tbl and H_PUT_TCE-hypercalls are
expected to update every single table in the list.

This adds kvm_spapr_tce_attach_iommu_group() helper to initialize
the data structs.

This puts the group when guest copy of TCE table is destroyed which
happens when TCE table fd is closed.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>