SPI Orion: Initialize stack variable to make gcc behave correctly
This enables SPI which in turn enables NAND SPI which in turn enables
/dev/mtd0 to read real MAC address from.

For whatever reason even if the code below goes via reading
"num-cs" from the device tree, the actual num_cs variable remains zero
but only sometime.

This simple fix fixes it for gcc (Debian 4.9.2-10) available on
Debian 8 at the time.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>