KVM: PPC: Add in-kernel acceleration for VFIO
This allows the host kernel to handle H_PUT_TCE, H_PUT_TCE_INDIRECT
and H_STUFF_TCE requests targeted an IOMMU TCE table used for VFIO
without passing them to user space which saves time on switching
to user space and back.

This adds H_PUT_TCE/H_PUT_TCE_INDIRECT/H_STUFF_TCE handlers to KVM.
KVM tries to handle a TCE request in the real mode, if failed
it passes the request to the virtual mode to complete the operation.
If it a virtual mode handler fails, the request is passed to
the user space; this is not expected to happen though.

To avoid dealing with page use counters (which is tricky in real mode),
this only accelerates SPAPR TCE IOMMU v2 clients which are required
to pre-register the userspace memory. The very first TCE request will
be handled in the VFIO SPAPR TCE driver anyway as the userspace view
of the TCE table (iommu_table::it_userspace) is not allocated till
the very first mapping happens and we cannot call vmalloc in real mode.

This adds new attribute - KVM_DEV_VFIO_GROUP_SET_SPAPR_TCE - to
the VFIO KVM device. It takes a VFIO group fd and SPAPR TCE table fd
and associates a physical IOMMU table with the SPAPR TCE table (which
is a guest view of the hardware IOMMU table). The iommu_table object
is referenced so we do not have to retrieve in real mode when hypercall
happens.

This does not implement the UNSET counterpart as there is no use for it -
once the acceleration is enabled, the existing userspace won't
disable it unless a VFIO container is detroyed so this adds necessary
cleanup to the KVM_DEV_VFIO_GROUP_DEL handler.

This uses the kvm->lock mutex to protect against a race between
the VFIO KVM device's kvm_vfio_destroy() and SPAPR TCE table fd's
release() callback.

This uses per recently introduced VFIO group notifiers to do VFIO KVM
device cleanup. The KVM_DEV_VFIO_GROUP_SET_SPAPR_TCE handler registers
a notifier which removed itself from the notifier chain when called
with kvm==NULL. Since a block notifier cannot remove itself when
it is executing, this uses RCU to do actual removal. To make this work,
the KVM holds an additional (to KVM_DEV_VFIO_GROUP_ADD) external user
reference to a group and also relies on
"vfio: Check for unregistered notifiers when group is actually released".
While it works in most cases, there is a problem with VFIO PCI hotunplug:
QEMU calls vfio_kvm_device_del_group() and vfio_disconnect_container()
and the VFIO_GROUP_UNSET_CONTAINER handler may fail if RCU handler
has not been called yet so group->container_users counter is still
non-zero; this is workarounded by retrying VFIO_GROUP_UNSET_CONTAINER
in QEMU (this needs to be fixed, suggestions are welcome).

As this creates a descriptor per IOMMU table-LIOBN couple (called
kvmppc_spapr_tce_iommu_table), it is possible to have several
descriptors with the same iommu_table (hardware IOMMU table) attached
to the same LIOBN; this is done to simplify the cleanup and can be
improved later (even though there does not seem to be a user for it
any time soon).

This advertises the new KVM_CAP_SPAPR_TCE_VFIO capability to the user
space.

This finally makes use of vfio_external_user_iommu_id() which was
introduced quite some time ago and was considered for removal.

Tests show that this patch increases transmission speed from 220MB/s
to 750..1020MB/s on 10Gb network (Chelsea CXGB3 10Gb ethernet card).

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
---
Changes:
v2:
* reworked to use new VFIO notifiers
* now same iommu_table may appear in the list several times, to be fixed later