powerpc/ioda/npu2: Call hot reset skiboot hook when disabling NPU2
Normally GPU RAM is onlined and offline by the deamon and the driver;
however in case of guest termination neither gets chance to offline
GPU RAM so we end up with a running machine which is still having memory
but accessing this memory produces hardware management interrupts (HMI)
which bring the host down.

The skiboot firmware has a hot reset handler which fences the GPU RAM
and makes accesses no-op instead of HMI:
https://github.com/open-power/skiboot/commit/fca2b2b839a67

This wires up this hot reset hook to vfio_pci_disable() via
pci_disable_device() which brings NPU2 in a safe mode and prevents HMIs.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
---
Changes:
v2:
* updated the commit log