powerpc/powernv: Detach npu struct from pnv_phb
The powernv PCI code stores NPU data in the pnv_phb struct reference to
which is stored in pci_controller::private_data.

We are going to have NPU2 support in pseries platform as well but it does
not store any private_data in in the pci_controller struct.

This adds a global list of NPUs so each platform can register and use
these in the same fashion.

This moves pnv_npu2_init() declaration out of the powernv folder so
later pseries could include it.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>