powerpc/powernv/ioda2: Reduce upper limit for DMA window size (retry)
We use PHB in mode1 which uses bit 59 to select a correct DMA window.
However there is mode2 which uses bits 59:55 and allows up to 32 DMA
windows per a PE.

Even though documentation does not clearly specify that, it seems that
the actual hardware does not support bits 59:55 even in mode1, in other
words we can create a window as big as 1<<58 but DMA simply won't work.

This reduces the upper limit from 59 to 55 bits to let the userspace know
about the hardware limits.

Fixes: ce57c6610cc2 "Merge branch 'topic/ppc-kvm' into next"
Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
---

The merge commit says  d3d4ffaae439 (the original of this one) was
propagated but it was not:

[vpl1 kernel]$ git s ce57c6610cc2:arch/powerpc/platforms/powernv/pci-ioda-tce.c | grep 'page_shift >= '
        if ((level_shift - 3) * levels + page_shift >= 60)

(cherry picked from commit 709da552a72888e1bbf7e0945a574f5566c0f782)
Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>