powerpc/powernv/ioda: Allocate indirect TCE levels of cached userspace addresses on demand
In order to support VFIO on the powernv platform, the IOMMU code maintains
2 TCE tables - one for the hardware with the host physical addresses and
another one with userspace addresses used for mapping; this is used to
mark pages dirty when corresponging TCEs are unmapped from the hardware
table.

a68bd1267b72 enabled on-demand allocation of the hardware table,
however due to oversight it missed the other table so it has still been
fully allocated at the boot time. This fixes the issue by allocating
a single level just like we do already for the hardware table.

Fixes: a68bd1267b72 "powerpc/powernv/ioda: Allocate indirect TCE levels on demand"
Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>