vfio_pci: Add NVIDIA GV100GL [Tesla V100 SXM2] [10de:1db1] subdriver
Some POWER9 chips come with special NVLink2 links which provide
cacheable memory access to the RAM physically located on NVIDIA GPU.
This memory is presented to a host via the device tree but remains
offline until the NVIDIA driver onlines it.

This exports this RAM to the userspace as a new region so
the NVIDIA driver in the guest can train these links and online GPU RAM.

This exports ATSD register to do whatever the driver wants to do on
the baremetal.

This does want KVM as the nest MMU needs to be told about LPID
(logical partition ID, a guest ID) for ATS.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>