KVM: PPC: Optimize clearing TCEs for sparse tables
The PPC VFIO IOMMU code manages a userspace addresses cache multilevel
table non-root levels of which are allocated on demand. Since
the memory allocation is problematic in real mode, we have 2 accessors
to the entrues - one which allocates the memory and is always expected
to return non-NULL and another which does not do allocations and can
return NULL; the latter is used for the real mode only.

Also, DMA windows can span to up to 55 bits of the address space and when
such windows are used, they have very sparse mappings utilizing only
a fraction of the window. And yet the SPAPR TCE IOMMU subdriver still has
to walk through all TCEs and unpin mapped memory.

In cases of such huge sparse windows we are likely to see multilevel
tables in use (basically a radix tree), which may not have the entire
tree allocated so we can take a shortcut here and skip some TCEs - for
this we look at the userspace addresses cache (which is a mirror of
the hardware TCE table) and  if it is missing some table leaves, then
the hardware table does not have them allocated either and therefore
does not require updating.

This helps a lot with sparse DMA windows, reducing the shutdown time from
about 3 minutes per 1 billion TCEs to a few seconds for 32GB sparse case
which is less than reset time for many devices. Just skipping the last
level seems to be good enough not to bother with skipping more TCEs
at once at higher levels (which is possible though, just requires
extention to the it_ops->useraddrptr() which does not feel worth).

As non-allocating accessor is used now in virtual mode as well, rename it
from IOMMU_TABLE_USERSPACE_ENTRY_RM (real mode) to _RO (read only).

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>