powerpc/powernv/npu: Add helper to map GPU to LPAR
In order to make ATS work and translate addresses for arbitrary
LPID and PID, we need to program an NPU with these.

This implements a helper to assign a GPU to LPAR and program the NPU
with a wildcard for PID and a helper to do clean-up. The helper takes
MSR (only DR/HV/PR/SF bits are allowed) to program them into NPU2 for
ATS checkout requests support.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>